segmentation_tool/requirements.txt
Flask
Pillow
numpy<2
